+++ CHANGELOG.md
@@ -11,6 +11,8 @@ Changelog
 
 - Allow splitting up "birthday" and "date" fields into separate fields with `day`, `month` and `year` index.
 - Improved algorithm for finding fields when integrating with Contact Form 7 or other custom forms.
+- Ninja Forms integration can now find automatically find name-fields.
+- Ninja Forms integration can now use `mc4wp-` prefixed admin labels.
 
 #### 3.0.10 - January 6, 2016